Convert Commentable to trait
