+++ grails-app/conf/application.groovy
@@ -1,1 +1,1 @@
-grails.commentable.poster.evaluator = { [id: 1, class: [name:'org.grails.comments.TestPoster']] }
+grails.commentable.poster.evaluator = { request.user }
