+++ src/main/groovy/grails/plugins/commentable/CommentableGrailsPlugin.groovy
@@ -48,91 +48,4 @@ then use the tag library and partial templates to integrate comments into your v
 			config.grails.commentable.poster.evaluator = { request.user }
 		}
 	}
-
-	def doWithDynamicMethods = { ctx ->
-		for(domainClass in grailsApplication.domainClasses) {
-			if(Commentable.class.isAssignableFrom(domainClass.clazz)) {
-				domainClass.clazz.metaClass {
-					'static' {
-						getRecentComments {->
-							def clazz = delegate
-							CommentLink.withCriteria {
-								projections { property "comment" }
-								eq 'type',GrailsNameUtils.getPropertyName(clazz)
-								maxResults 5
-								cache true
-								comment {
-									order "dateCreated", "desc"
-								}
-							}
-						}
-					}
-					
-					addComment { poster, String text ->
-						if(delegate.id == null) throw new CommentException("You must save the entity [${delegate}] before calling addComment")
-						
-						def posterClass = poster.class.name
-						def i = posterClass.indexOf('_$$_javassist')
-						if(i>-1)
-							posterClass = posterClass[0..i-1]
-						
-						def c = new Comment(body:text, posterId:poster.id, posterClass:posterClass)
-						if(!c.validate()) {
-							throw new CommentException("Cannot create comment for arguments [$poster, $text], they are invalid.")
-						}
-						c.save()
-						def link = new CommentLink(comment:c, commentRef:delegate.id, type:GrailsNameUtils.getPropertyName(delegate.class))
-						link.save()
-						try {
-							delegate.onAddComment(c)
-						} catch (MissingMethodException e) {}
-						return delegate
-					}
-
-					getComments = {->
-						def instance = delegate
-						if(instance.id != null) {
-							CommentLink.withCriteria {
-								projections {
-									property "comment"
-								}
-								eq "commentRef", instance.id
-								eq 'type', GrailsNameUtils.getPropertyName(instance.class)
-								cache true
-							}
-						} else {
-							return Collections.EMPTY_LIST
-						}
-					}
-					
-					getTotalComments = {->
-						def instance = delegate
-						if(instance.id != null) {
-							CommentLink.createCriteria().get {
-								projections {
-									rowCount()
-								}
-								eq "commentRef", instance.id
-								eq 'type', GrailsNameUtils.getPropertyName(instance.class)
-								cache true
-							}
-						} else {
-							return 0
-						}
-					}
-					
-					removeComment { Comment c ->
-						CommentLink.findAllByComment(c)*.delete()
-						c.delete(flush:true) // cascades deletes to links
-					}
-					
-					removeComment { Long id ->
-						def c = Comment.get(id)
-						if(c) removeComment(c)
-					}
-				}
-			}
-		}
-	}
-
 }
